LSR-291 adapt test-app

+++ android/samples/test-app/src/main/java/com/liferay/mobile/screens/testapp/DDLFormActivity.java
@@ -130,9 +130,9 @@ public class DDLFormActivity extends ThemeActivity implements DDLFormListener {
 
 	private void initScreenletFromIntent(Intent intent) {
 		if (intent.hasExtra("recordId")) {
-			_screenlet.setRecordId(intent.getIntExtra("recordId", 0));
-			_screenlet.setRecordSetId(intent.getIntExtra("recordSetId", 0));
-			_screenlet.setStructureId(intent.getIntExtra("structureId", 0));
+			_screenlet.setRecordId(intent.getLongExtra("recordId", 0));
+			_screenlet.setRecordSetId(intent.getLongExtra("recordSetId", 0));
+			_screenlet.setStructureId(intent.getLongExtra("structureId", 0));
 		}
 	}
